+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 0.8.3
 
 - Bugfix: `allow_substates` did not consider enum states
+- Feature: Nested enums can now be passed in a dict as `children` with `initial` parameter
 
 ## 0.8.2 (June 2020)