+++ Changelog.md
@@ -7,6 +7,7 @@
 - Bugfix #449: get_triggers/get_transitions did not return nested triggers correctly (thanks @alexandretanem)
 - Feature #452: Improve handling of label attributes in custom diagram states and `TransitionGraphSupport` (thanks @badiku)
 - Bugfix #456: Prevent parents from overriding (falsy) results of their children's events (thanks @alexandretanem)
+- Bugfix #458: Entering the same state caused key errors when transition was defined on a parent (thanks @matlom)
 
 ## 0.8.2 (June 2020)